requirements.txt
Flask
requests
python-whois